Refactor IBA -- break up unwieldy imagebufalgo.cpp. Move all high-quality filtered transformations/warps to imagebufalgo_xform.cpp. Move all drawing and visible annotation to imagebufalgo_draw.cpp.
No actual code or functionality has changed, just moving things from file
to file for better organization and faster recompilation.
