+++ src/libOpenImageIO/CMakeLists.txt
@@ -16,7 +16,9 @@ set (libOpenImageIO_srcs exif.cpp formatspec.cpp imagebuf.cpp
                           imagebufalgo_compare.cpp
                           imagebufalgo_copy.cpp
                           imagebufalgo_deep.cpp
+                          imagebufalgo_draw.cpp
                           imagebufalgo_pixelmath.cpp
+                          imagebufalgo_xform.cpp
                           imagebufalgo_yee.cpp imagebufalgo_opencv.cpp
                           maketexture.cpp
                           ../libutil/argparse.cpp
